migrate card molecule test to testing-library
